functions/local_test/requirements.txt
firebase_functions~=0.1.0
pydantic
langchain
pypdf
openai==1.3.3
beautifulsoup4
requests
scikit-learn
numpy
python-dotenv
